Move source files from utility to datastructs
Also, remove PrefilledMirroredRingBuffer
