+++ include/uit/conduit/proc/inlet/put=dropping+type=trivial/impl/TrivialRingImmediateSendDuct.hpp
@@ -15,7 +15,7 @@
 #include "../../../../../mpi/Request.hpp"
 #include "../../../../../utility/CircularIndex.hpp"
 #include "../../../../../utility/print_utils.hpp"
-#include "../../../../../utility/RingBuffer.hpp"
+#include "../../../../../datastructs/RingBuffer.hpp"
 
 #include "../../../../InterProcAddress.hpp"
