+++ include/uit/conduit/proc/outlet/get=stepping+type=trivial/t::RingIrecvDuct.hpp
@@ -12,9 +12,9 @@
 
 #include "../../../../debug/err_audit.hpp"
 #include "../../../../mpi/mpi_utils.hpp"
-#include "../../../../utility/SiftingArray.hpp"
+#include "../../../../datastructs/SiftingArray.hpp"
 #include "../../../../utility/print_utils.hpp"
-#include "../../../../utility/RingBuffer.hpp"
+#include "../../../../datastructs/RingBuffer.hpp"
 
 #include "../../../InterProcAddress.hpp"
