+++ tests/Makefile
@@ -5,6 +5,7 @@ TARGET_NAMES += concurrent
 TARGET_NAMES += conduit
 TARGET_NAMES += containers
 TARGET_NAMES += countdown
+TARGET_NAMES += datastructs
 TARGET_NAMES += debug
 TARGET_NAMES += distributed
 TARGET_NAMES += math
